pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "niuma-bot"
version = "0.1.0"
description = "Teams chat bot powered by Claude Code"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
    "aiosqlite>=0.19.0",
    "markdown>=3.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
]

[project.scripts]
niuma = "niuma.main:cli_entry"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
